Improved: moved timeZone modal from views to component
